+++ src/main/java/net/frozenblock/wilderwild/tag/WWBiomeTags.java
@@ -26,7 +26,7 @@ import org.jetbrains.annotations.NotNull;
 public final class WWBiomeTags {
 	public static final TagKey<Biome> STRAYS_CAN_SPAWN_UNDERGROUND = bind("entity/spawn/underground_stray");
 	public static final TagKey<Biome> HAS_BUTTERFLY = bind("entity/spawn/butterfly");
-	public static final TagKey<Biome> HAS_BUTTERFLY_COMMON = bind("entity/spawn/butterfly_common");
+	public static final TagKey<Biome> BUTTERFLY_COMMON_SPAWN = bind("entity/spawn/butterfly_common_spawn");
 	public static final TagKey<Biome> HAS_FIREFLY = bind("entity/spawn/firefly");
 	public static final TagKey<Biome> HAS_JELLYFISH = bind("entity/spawn/jellyfish");
 	public static final TagKey<Biome> JELLYFISH_COMMON_SPAWN = bind("entity/spawn/jellyfish_common_spawn");