+++ CHANGELOG.md
@@ -8,4 +8,6 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
     - The `wilderwild:entity/spawn/jellyfish` biome tag now dictates all biomes Jellyfish can spawn in.
     - The `wilderwild:entity/spawn/jellyfish_special_spawn` biome tag has been replaced with the `wilderwild:entity/spawn/jellyfish_common_spawn` biome tag.
       - This tag now dictates which biomes (of the ones they can spawn in) Jellyfish can spawn more frequently and closer together in.
-- The `wilderwild:entity/spawn/common_butterfly` biome tag has been renamed to `wilderwild:entity/spawn/butterfly_common`.
+    - The `wilderwild:entity/spawn/butterfly` biome tag now dictates all biomes Butterflies can spawn in.
+    - The `wilderwild:entity/spawn/common_butterfly` biome tag has been replaced with the `wilderwild:entity/spawn/butterfly_common_spawn` biome tag.
+        - This tag now dictates which biomes (of the ones they can spawn in) Butterflies can spawn more frequently and closer together in.